Allow passing aliases to --index
Currently if you pass an alias name to --index, elyzer throws an
AnalyzerNotFound exception, which is a bit confusing because the
analyzer does exist.

This is because even though the get settings API supports passing an
alias, the response is always keyed by index name, so elyzer hits a
KeyError when looking for the alias name.

In the case where an alias points to a single index, we can just take
whatever is returned instead of looking up the name in the response.

Passing an alias or wildcard matching multiple indexes still throws an
error - in this case I think the user should be more specific.